Ensure clone supplies subclass constructor requirements
