+++ core/units/__init__.py
@@ -1,0 +1,24 @@
+from .cog_unit import CogUnit, EmitterUnit, ProcessorUnit, SensorUnit
+from .constants import (
+    ENABLE_MINI_LEARN,
+    FOLLOW_INPUT_DEVICE,
+    MAX_OUTPUT_DIM,
+    ROLE_SPLIT_RULE,
+    SPLIT_HI_ES_TABLE,
+    SPLIT_HI_P_TABLE,
+    TOL_FRAC_SPLIT,
+)
+
+__all__ = [
+    "CogUnit",
+    "SensorUnit",
+    "ProcessorUnit",
+    "EmitterUnit",
+    "ENABLE_MINI_LEARN",
+    "FOLLOW_INPUT_DEVICE",
+    "MAX_OUTPUT_DIM",
+    "ROLE_SPLIT_RULE",
+    "SPLIT_HI_ES_TABLE",
+    "SPLIT_HI_P_TABLE",
+    "TOL_FRAC_SPLIT",
+]
